LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-grid-003-expected.txt


PASS .target > * 1
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<div data-offset-x="93"><span></span></div>
offsetLeft expected 93 but got 92
PASS .target > * 4
FAIL .target > * 5 assert_equals:
<div data-offset-x="50"><span></span></div>
offsetLeft expected 50 but got 25
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<div data-offset-x="108"><span></span></div>
offsetLeft expected 108 but got 87
PASS .target > * 8
PASS .target > * 9
PASS .target > * 10
FAIL .target > * 11 assert_equals:
<div data-offset-x="88"><span></span></div>
offsetLeft expected 88 but got 87
PASS .target > * 12
